type change to follow strictly the api

+++ tests/PixTest.hh
@@ -36,7 +36,7 @@ public:
   virtual std::string getTestTip() {return fTestTip;}
 
   int pixelThreshold(std::string dac, int ntrig, int dacmin, int dacmax);
-  std::vector<TH2D*> efficiencyMaps(std::string name, int ntrig = 10); 
+  std::vector<TH2D*> efficiencyMaps(std::string name, uint16_t ntrig = 10); 
   std::vector<TH1*> scurveMaps(std::string dac, std::string name, int ntrig = 10, int daclo = 0, int dachi = 255, int result = 3); 
   std::vector<TH1*> thrMaps(std::string dac, std::string name, int ntrig = 10);
